+++ TaxareProject/SeguridadSocial.cs
@@ -15,7 +15,7 @@ namespace TaxareProject
     public partial class SeguridadSocial
     {
         public int id { get; set; }
-        public int id_conductor { get; set; }
+        public long id_conductor { get; set; }
         public System.DateTime pago_anterior { get; set; }
         public System.DateTime pago_siguiente { get; set; }
         public double valor { get; set; }
